+++ packages/atomic/src/components/atomic-smart-snippet/atomic-smart-snippet.tsx
@@ -1,4 +1,4 @@
-import {Component, h, Prop, State} from '@stencil/core';
+import {Component, h, Prop, State, Element} from '@stencil/core';
 import {
   InitializableComponent,
   InitializeBindings,
@@ -49,12 +49,26 @@ export class AtomicSmartSnippet implements InitializableComponent {
   @State()
   public smartSnippetState!: SmartSnippetState;
   public error!: Error;
+  @Element() public host!: HTMLElement;
   private id = randomID();
 
   /**
    * The [heading level](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Heading_Elements) to use for the question at the top of the snippet, from 1 to 5.
    *
    * We recommend setting this property in order to improve accessibility.
+   *
+   * You can style the snippet by inserting a template element like this:
+   * ```html
+   * <atomic-smart-snippet>
+   *   <template>
+   *     <style>
+   *       b {
+   *         color: blue;
+   *       }
+   *     </style>
+   *   </template>
+   * </atomic-smart-snippet>
+   * ```
    */
   @Prop({reflect: true}) public headingLevel = 0;
 
@@ -66,9 +80,32 @@ export class AtomicSmartSnippet implements InitializableComponent {
    * When the answer is partly hidden, how much of its height (in pixels) should be visible.
    */
   @Prop({reflect: true}) collapsedHeight = 180;
+  /**
+   * Sets the style of the snippet.
+   *
+   * Example:
+   * ```ts
+   * smartSnippet.snippetStyle = `
+   *   b {
+   *     color: blue;
+   *   }
+   * `;
+   * ```
+   */
+  @Prop({reflect: true}) snippetStyle?: string;
 
   public initialize() {
     this.smartSnippet = buildSmartSnippet(this.bindings.engine);
+  }
+
+  private get style() {
+    const styleTag = this.host
+      .querySelector('template')
+      ?.content.querySelector('style');
+    if (!styleTag) {
+      return this.snippetStyle;
+    }
+    return styleTag.innerHTML;
   }
 
   private renderQuestion() {
@@ -90,6 +127,7 @@ export class AtomicSmartSnippet implements InitializableComponent {
         part="body"
         maximumHeight={this.maximumHeight}
         collapsedHeight={this.collapsedHeight}
+        snippetStyle={this.style}
       ></atomic-smart-snippet-expandable-answer>
     );
   }
@@ -137,7 +175,7 @@ export class AtomicSmartSnippet implements InitializableComponent {
     );
   }
 
-  public renderFeedbackBanner() {
+  private renderFeedbackBanner() {
     return (
       <SmartSnippetFeedbackBanner
         i18n={this.bindings.i18n}