+++ packages/headless/src/api/analytics/analytics.ts
@@ -5,6 +5,7 @@ import {
   CoveoAnalyticsClient,
   AnalyticsClientSendEventHook,
   CaseAssistClient,
+  CaseAssistClientProvider,
 } from 'coveo.analytics';
 import {Logger} from 'pino';
 import {getQueryInitialState} from '../../features/query/query-state';
@@ -43,7 +44,8 @@ export type StateNeededByAnalyticsProvider = ConfigurationSection &
   >;
 
 export type StateNeededByCaseAssistAnalytics = ConfigurationSection &
-  Partial<CaseAssistConfigurationSection>;
+  Partial<CaseAssistConfigurationSection> &
+  Partial<SearchHubSection>;
 
 export class AnalyticsProvider implements SearchPageClientProvider {
   constructor(private state: StateNeededByAnalyticsProvider) {}
@@ -139,6 +141,14 @@ export class AnalyticsProvider implements SearchPageClientProvider {
   }
 }
 
+export class CaseAssistAnalyticsProvider implements CaseAssistClientProvider {
+  constructor(private state: StateNeededByCaseAssistAnalytics) {}
+
+  public getOriginLevel1() {
+    return this.state.searchHub || getSearchHubInitialState();
+  }
+}
+
 interface ConfigureAnalyticsOptions {
   state: StateNeededByAnalyticsProvider;
   logger: Logger;
@@ -212,6 +222,7 @@ interface ConfigureCaseAssistAnalyticsOptions {
   logger: Logger;
   analyticsClientMiddleware?: AnalyticsClientSendEventHook;
   preprocessRequest?: PreprocessRequest;
+  provider?: CaseAssistClientProvider;
 }
 
 export const configureCaseAssistAnalytics = ({
@@ -219,33 +230,37 @@ export const configureCaseAssistAnalytics = ({
   state,
   analyticsClientMiddleware = (_, p) => p,
   preprocessRequest,
+  provider = new CaseAssistAnalyticsProvider(state),
 }: ConfigureCaseAssistAnalyticsOptions) => {
   const token = state.configuration.accessToken;
   const endpoint = state.configuration.analytics.apiBaseUrl;
   const runtimeEnvironment = state.configuration.analytics.runtimeEnvironment;
   const enableAnalytics = state.configuration.analytics.enabled;
-  const client = new CaseAssistClient({
-    enableAnalytics,
-    token,
-    endpoint,
-    runtimeEnvironment,
-    preprocessRequest,
-    beforeSendHooks: [
-      analyticsClientMiddleware,
-      (type, payload) => {
-        logger.info(
-          {
-            ...payload,
-            type,
-            endpoint,
-            token,
-          },
-          'Analytics request'
-        );
-        return payload;
-      },
-    ],
-  });
+  const client = new CaseAssistClient(
+    {
+      enableAnalytics,
+      token,
+      endpoint,
+      runtimeEnvironment,
+      preprocessRequest,
+      beforeSendHooks: [
+        analyticsClientMiddleware,
+        (type, payload) => {
+          logger.info(
+            {
+              ...payload,
+              type,
+              endpoint,
+              token,
+            },
+            'Analytics request'
+          );
+          return payload;
+        },
+      ],
+    },
+    provider
+  );
 
   if (!enableAnalytics) {
     client.disable();