+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.11...@coveo/atomic-angular@0.8.12) (2022-04-07)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.10...@coveo/atomic-angular@0.8.11) (2022-04-06)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.10](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.9...@coveo/atomic-angular@0.8.10) (2022-04-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.8...@coveo/atomic-angular@0.8.9) (2022-04-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [0.8.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.7...@coveo/atomic-angular@0.8.8) (2022-04-04)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [0.8.7](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@0.8.6...@coveo/atomic-angular@0.8.7) (2022-04-01)
 
 **Note:** Version bump only for package @coveo/atomic-angular